+++ packages/newcar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # newcar
 
+## 0.7.1
+
+### Patch Changes
+
+- 2554e8b: fix: The setup function runs before the scene was checked out
+- Updated dependencies [2554e8b]
+  - @newcar/core@0.7.1
+  - @newcar/basic@0.7.1
+  - @newcar/recorder@0.7.1
+  - @newcar/utils@0.7.1
+
 ## 0.7.0
 
 ### Minor Changes