+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @newcar/utils
 
+## 0.7.1
+
+### Patch Changes
+
+- 2554e8b: fix: The setup function runs before the scene was checked out
+
 ## 0.7.0
 
 ### Minor Changes